+++ CHANGELOG.md
@@ -18,13 +18,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Restore default idempotency of `PUT /api/nodes/{id}`
 - `DELETE /api/overlays/{name}?force=true` can delete overlays that are in use
 - `warewulfd` overlay autobuild rebuilds overlays after node discovery. #1468
-### Fixed
-
 - Improved netplan support. #1873
 
 ### Fixed
 
 - Fixed a bug when cloning an overlay to site when parent is missing
+- Fixed `wwctl upgrade nodes` to properly handle kernel argument lists. #1938
 
 ## v4.6.2, 2025-07-09